requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
smolagents[e2b]
python-dotenv
pyyaml